requirements.txt
openai>=1.0.0
pyyaml>=6.0.0
requests>=2.31.0
speechrecognition>=3.10.0
pyttsx3>=2.90
